Refactor content type creation to sidebar
